# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: pharmacies
    description: Pharmacy management
  - name: patients
    description: Patient management
  - name: ocr-logs
    description: AI/OCR scan logs
  - name: orders
    description: Order management
  - name: settings
    description: Platform settings
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /pharmacies:
    get:
      operationId: listPharmacies
      tags: [pharmacies]
      summary: List all pharmacies
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of pharmacies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pharmacy"
    post:
      operationId: createPharmacy
      tags: [pharmacies]
      summary: Create a pharmacy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePharmacyBody"
      responses:
        "201":
          description: Created pharmacy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pharmacy"

  /pharmacies/{id}:
    get:
      operationId: getPharmacy
      tags: [pharmacies]
      summary: Get a pharmacy by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pharmacy details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pharmacy"
        "404":
          description: Not found

  /pharmacies/{id}/status:
    patch:
      operationId: updatePharmacyStatus
      tags: [pharmacies]
      summary: Update pharmacy status (approve or reject)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePharmacyStatusBody"
      responses:
        "200":
          description: Updated pharmacy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pharmacy"
        "404":
          description: Not found

  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List all patients
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, inactive]
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patient"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Create a patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePatientBody"
      responses:
        "201":
          description: Created patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"

  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get a patient by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "404":
          description: Not found
    patch:
      operationId: updatePatient
      tags: [patients]
      summary: Update a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePatientBody"
      responses:
        "200":
          description: Updated patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "404":
          description: Not found

  /ocr-logs:
    get:
      operationId: listOcrLogs
      tags: [ocr-logs]
      summary: List AI/OCR scan logs
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [success, manual_fallback, failed]
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of OCR logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OcrLog"

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [processing, shipped, delivered, cancelled]
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found

  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get platform settings
      responses:
        "200":
          description: Platform settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update platform settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsBody"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Pharmacy:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        registrationNumber:
          type: string
        location:
          type: string
        contactEmail:
          type: string
        phone:
          type: string
        licenseExpiry:
          type: string
        username:
          type: string
        passwordHash:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        dateApplied:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - registrationNumber
        - location
        - contactEmail
        - phone
        - licenseExpiry
        - status
        - dateApplied
        - username
        - passwordHash
        - createdAt
        - updatedAt

    CreatePharmacyBody:
      type: object
      properties:
        name:
          type: string
        registrationNumber:
          type: string
        location:
          type: string
        contactEmail:
          type: string
        phone:
          type: string
        licenseExpiry:
          type: string
        dateApplied:
          type: string
        username:
          type: string
        password:
          type: string
      required:
        - name
        - registrationNumber
        - location
        - contactEmail
        - phone
        - licenseExpiry
        - dateApplied
        - username
        - password

    UpdatePharmacyStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [approved, rejected]
      required:
        - status

    Patient:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        address:
          type: string
        prescriptionCount:
          type: integer
        lastVisit:
          type: string
        status:
          type: string
          enum: [active, inactive]
        username:
          type: string
        passwordHash:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - email
        - phone
        - dateOfBirth
        - address
        - prescriptionCount
        - lastVisit
        - status
        - username
        - passwordHash
        - createdAt
        - updatedAt

    CreatePatientBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        address:
          type: string
        lastVisit:
          type: string
        username:
          type: string
        password:
          type: string
      required:
        - name
        - email
        - phone
        - dateOfBirth
        - address
        - lastVisit
        - username
        - password

    UpdatePatientBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        status:
          type: string
          enum: [active, inactive]
        lastVisit:
          type: string

    OcrLog:
      type: object
      properties:
        id:
          type: integer
        prescriptionId:
          type: string
        medicationName:
          type: string
        pharmacyName:
          type: string
        patientName:
          type: string
        status:
          type: string
          enum: [success, manual_fallback, failed]
        confidence:
          type: number
        errorReason:
          type: ["string", "null"]
        scannedAt:
          type: string
        createdAt:
          type: string
      required:
        - id
        - prescriptionId
        - medicationName
        - pharmacyName
        - patientName
        - status
        - confidence
        - errorReason
        - scannedAt
        - createdAt

    Order:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: string
        patientName:
          type: string
        pharmacyName:
          type: string
        medications:
          type: string
        status:
          type: string
          enum: [processing, shipped, delivered, cancelled]
        total:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - orderId
        - patientName
        - pharmacyName
        - medications
        - status
        - total
        - createdAt
        - updatedAt

    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [processing, shipped, delivered, cancelled]
      required:
        - status

    Settings:
      type: object
      properties:
        id:
          type: integer
        platformName:
          type: string
        supportEmail:
          type: string
        ocrConfidenceThreshold:
          type: number
        autoApprovePharmacies:
          type: boolean
        maxPrescriptionsPerDay:
          type: integer
        maintenanceMode:
          type: boolean
        updatedAt:
          type: string
      required:
        - id
        - platformName
        - supportEmail
        - ocrConfidenceThreshold
        - autoApprovePharmacies
        - maxPrescriptionsPerDay
        - maintenanceMode
        - updatedAt

    UpdateSettingsBody:
      type: object
      properties:
        platformName:
          type: string
        supportEmail:
          type: string
        ocrConfidenceThreshold:
          type: number
        autoApprovePharmacies:
          type: boolean
        maxPrescriptionsPerDay:
          type: integer
        maintenanceMode:
          type: boolean