# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all pharmacies
 */
export const ListPharmaciesQueryParams = zod.object({
  status: zod.enum(["pending", "approved", "rejected"]).optional(),
  search: zod.coerce.string().optional(),
});

export const ListPharmaciesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  registrationNumber: zod.string(),
  location: zod.string(),
  contactEmail: zod.string(),
  phone: zod.string(),
  licenseExpiry: zod.string(),
  username: zod.string(),
  passwordHash: zod.string(),
  status: zod.enum(["pending", "approved", "rejected"]),
  dateApplied: zod.string(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListPharmaciesResponse = zod.array(ListPharmaciesResponseItem);

/**
 * @summary Create a pharmacy
 */
export const CreatePharmacyBody = zod.object({
  name: zod.string(),
  registrationNumber: zod.string(),
  location: zod.string(),
  contactEmail: zod.string(),
  phone: zod.string(),
  licenseExpiry: zod.string(),
  dateApplied: zod.string(),
  username: zod.string(),
  password: zod.string(),
});

/**
 * @summary Get a pharmacy by ID
 */
export const GetPharmacyParams = zod.object({
  id: zod.coerce.number(),
});

export const GetPharmacyResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  registrationNumber: zod.string(),
  location: zod.string(),
  contactEmail: zod.string(),
  phone: zod.string(),
  licenseExpiry: zod.string(),
  username: zod.string(),
  passwordHash: zod.string(),
  status: zod.enum(["pending", "approved", "rejected"]),
  dateApplied: zod.string(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update pharmacy status (approve or reject)
 */
export const UpdatePharmacyStatusParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdatePharmacyStatusBody = zod.object({
  status: zod.enum(["approved", "rejected"]),
});

export const UpdatePharmacyStatusResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  registrationNumber: zod.string(),
  location: zod.string(),
  contactEmail: zod.string(),
  phone: zod.string(),
  licenseExpiry: zod.string(),
  username: zod.string(),
  passwordHash: zod.string(),
  status: zod.enum(["pending", "approved", "rejected"]),
  dateApplied: zod.string(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary List all patients
 */
export const ListPatientsQueryParams = zod.object({
  status: zod.enum(["active", "inactive"]).optional(),
  search: zod.coerce.string().optional(),
});

export const ListPatientsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  dateOfBirth: zod.string(),
  address: zod.string(),
  prescriptionCount: zod.number(),
  lastVisit: zod.string(),
  username: zod.string(),
  passwordHash: zod.string(),
  status: zod.enum(["active", "inactive"]),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListPatientsResponse = zod.array(ListPatientsResponseItem);

/**
 * @summary Create a patient
 */
export const CreatePatientBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  dateOfBirth: zod.string(),
  address: zod.string(),
  lastVisit: zod.string(),
  username: zod.string(),
  password: zod.string(),
});

/**
 * @summary Get a patient by ID
 */
export const GetPatientParams = zod.object({
  id: zod.coerce.number(),
});

export const GetPatientResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  dateOfBirth: zod.string(),
  address: zod.string(),
  prescriptionCount: zod.number(),
  lastVisit: zod.string(),
  username: zod.string(),
  passwordHash: zod.string(),
  status: zod.enum(["active", "inactive"]),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update a patient
 */
export const UpdatePatientParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdatePatientBody = zod.object({
  name: zod.string().optional(),
  email: zod.string().optional(),
  phone: zod.string().optional(),
  address: zod.string().optional(),
  status: zod.enum(["active", "inactive"]).optional(),
  lastVisit: zod.string().optional(),
});

export const UpdatePatientResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  dateOfBirth: zod.string(),
  address: zod.string(),
  prescriptionCount: zod.number(),
  lastVisit: zod.string(),
  username: zod.string(),
  passwordHash: zod.string(),
  status: zod.enum(["active", "inactive"]),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary List AI/OCR scan logs
 */
export const ListOcrLogsQueryParams = zod.object({
  status: zod.enum(["success", "manual_fallback", "failed"]).optional(),
  search: zod.coerce.string().optional(),
});

export const ListOcrLogsResponseItem = zod.object({
  id: zod.number(),
  prescriptionId: zod.string(),
  medicationName: zod.string(),
  pharmacyName: zod.string(),
  patientName: zod.string(),
  status: zod.enum(["success", "manual_fallback", "failed"]),
  confidence: zod.number(),
  errorReason: zod.string().nullable(),
  scannedAt: zod.string(),
  createdAt: zod.string(),
});
export const ListOcrLogsResponse = zod.array(ListOcrLogsResponseItem);

/**
 * @summary List all orders
 */
export const ListOrdersQueryParams = zod.object({
  status: zod
    .enum(["processing", "shipped", "delivered", "cancelled"])
    .optional(),
  search: zod.coerce.string().optional(),
});

export const ListOrdersResponseItem = zod.object({
  id: zod.number(),
  orderId: zod.string(),
  patientName: zod.string(),
  pharmacyName: zod.string(),
  medications: zod.string(),
  status: zod.enum(["processing", "shipped", "delivered", "cancelled"]),
  total: zod.number(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListOrdersResponse = zod.array(ListOrdersResponseItem);

/**
 * @summary Update order status
 */
export const UpdateOrderStatusParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateOrderStatusBody = zod.object({
  status: zod.enum(["processing", "shipped", "delivered", "cancelled"]),
});

export const UpdateOrderStatusResponse = zod.object({
  id: zod.number(),
  orderId: zod.string(),
  patientName: zod.string(),
  pharmacyName: zod.string(),
  medications: zod.string(),
  status: zod.enum(["processing", "shipped", "delivered", "cancelled"]),
  total: zod.number(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Get platform settings
 */
export const GetSettingsResponse = zod.object({
  id: zod.number(),
  platformName: zod.string(),
  supportEmail: zod.string(),
  ocrConfidenceThreshold: zod.number(),
  autoApprovePharmacies: zod.boolean(),
  maxPrescriptionsPerDay: zod.number(),
  maintenanceMode: zod.boolean(),
  updatedAt: zod.string(),
});

/**
 * @summary Update platform settings
 */
export const UpdateSettingsBody = zod.object({
  platformName: zod.string().optional(),
  supportEmail: zod.string().optional(),
  ocrConfidenceThreshold: zod.number().optional(),
  autoApprovePharmacies: zod.boolean().optional(),
  maxPrescriptionsPerDay: zod.number().optional(),
  maintenanceMode: zod.boolean().optional(),
});

export const UpdateSettingsResponse = zod.object({
  id: zod.number(),
  platformName: zod.string(),
  supportEmail: zod.string(),
  ocrConfidenceThreshold: zod.number(),
  autoApprovePharmacies: zod.boolean(),
  maxPrescriptionsPerDay: zod.number(),
  maintenanceMode: zod.boolean(),
  updatedAt: zod.string(),
});