# lib/api-zod/src/generated/types/createPatientBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePatientBody {
  name: string;
  email: string;
  phone: string;
  dateOfBirth: string;
  address: string;
  lastVisit: string;
}