# lib/api-zod/src/generated/types/ocrLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OcrLogStatus } from "./ocrLogStatus";

export interface OcrLog {
  id: number;
  prescriptionId: string;
  medicationName: string;
  pharmacyName: string;
  patientName: string;
  status: OcrLogStatus;
  confidence: number;
  /** @nullable */
  errorReason: string | null;
  scannedAt: string;
  createdAt: string;
}