# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PatientStatus } from "./patientStatus";

export interface Patient {
  id: number;
  name: string;
  email: string;
  phone: string;
  dateOfBirth: string;
  address: string;
  prescriptionCount: number;
  lastVisit: string;
  status: PatientStatus;
  createdAt: string;
  updatedAt: string;
}