# lib/api-zod/src/generated/types/pharmacy.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PharmacyStatus } from "./pharmacyStatus";

export interface Pharmacy {
  id: number;
  name: string;
  registrationNumber: string;
  location: string;
  contactEmail: string;
  phone: string;
  licenseExpiry: string;
  status: PharmacyStatus;
  dateApplied: string;
  createdAt: string;
  updatedAt: string;
}