# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Settings {
  id: number;
  platformName: string;
  supportEmail: string;
  ocrConfidenceThreshold: number;
  autoApprovePharmacies: boolean;
  maxPrescriptionsPerDay: number;
  maintenanceMode: boolean;
  updatedAt: string;
}