# lib/api-zod/src/generated/types/updatePatientBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdatePatientBodyStatus } from "./updatePatientBodyStatus";

export interface UpdatePatientBody {
  name?: string;
  email?: string;
  phone?: string;
  address?: string;
  status?: UpdatePatientBodyStatus;
  lastVisit?: string;
}