# lib/api-zod/src/generated/types/updateSettingsBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateSettingsBody {
  platformName?: string;
  supportEmail?: string;
  ocrConfidenceThreshold?: number;
  autoApprovePharmacies?: boolean;
  maxPrescriptionsPerDay?: number;
  maintenanceMode?: boolean;
}